chore(direct-message-list): move file to messages directory
Ref #227
