+++ apps/mull-ui/src/app/app.tsx
@@ -14,13 +14,13 @@ import { BotNavBar, PrivateRoute, SubNavBar, TopNavBar } from './components';
 import { UserProvider } from './context/user.context';
 import NotFoundPage from './pages/404/not-found-page';
 import CreateEventPage from './pages/create-event/create-event';
-import DirectMessageListPage from './pages/direct-message/direct-message-list/direct-message-list';
 import EventPage from './pages/event-page/event-page';
 import DiscoverPage from './pages/home/discover/discover-page';
 import MyEventsPage from './pages/home/my-events/my-events';
 import UpcomingPage from './pages/home/upcoming/upcoming';
 import LoginPage from './pages/login/login';
-import { EventMessageList } from './pages/messages/event-messages/event-message-list';
+import DirectMessageListPage from './pages/messages/direct-message/direct-message-list/direct-message-list';
+import EventMessageList from './pages/messages/event-messages/event-message-list';
 import AddFriendsPage from './pages/profile/add-friends/add-friends';
 import EditProfilePage from './pages/profile/edit-profile/edit-profile';
 import MyFriends from './pages/profile/myFriends-profile/myFriends-profile';
