+++ apps/mull-ui/src/app/pages/messages/direct-message/direct-message-list/direct-message-list.scss
@@ -1,4 +1,4 @@
-@import '../../../../variables.scss';
+@import '../../../../../variables.scss';
 
 .direct-messages-container {
   margin-top: 2rem;
