+++ apps/mull-ui/src/app/pages/messages/direct-message/direct-message-list/direct-message-list.spec.tsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import { createMemoryHistory } from 'history';
 import React from 'react';
 import { Router } from 'react-router-dom';
-import DirectMessagePage from './direct-message';
+import DirectMessageListPage from './direct-message-list';
 
 describe('DirectMessage', () => {
   it('should render successfully', () => {
@@ -11,7 +11,7 @@ describe('DirectMessage', () => {
     const { baseElement } = render(
       <MockedProvider>
         <Router history={history}>
-          <DirectMessagePage />
+          <DirectMessageListPage history={history} />
         </Router>
       </MockedProvider>
     );
